initial module lint refactoring
